Moved unit tests for pointer-values to a separate translation unit
