+++ tests/CMakeLists.txt
@@ -3,6 +3,29 @@ cmake_minimum_required (VERSION 3.2)
 option(SQLITE_ORM_OMITS_CODECVT "Omits codec testing" OFF)
 
 add_executable(unit_tests
+    static_tests/is_printable.cpp
+    static_tests/is_bindable.cpp
+    static_tests/iterator_t.cpp
+    static_tests/arithmetic_operators_result_type.cpp
+    static_tests/tuple_conc.cpp
+    static_tests/node_tuple.cpp
+    static_tests/bindable_filter.cpp
+    static_tests/count_tuple.cpp
+    static_tests/member_traits_tests.cpp
+    static_tests/select_return_type.cpp
+    static_tests/column.cpp
+    static_tests/same_or_void.cpp
+    static_tests/foreign_key.cpp
+    static_tests/tuple_filter.cpp
+    static_tests/storage_traits.cpp
+    static_tests/tuple_helper.cpp
+    static_tests/function_static_tests.cpp
+    static_tests/aggregate_function_return_types.cpp
+    static_tests/core_function_return_types.cpp
+    static_tests/column_result_t.cpp
+    static_tests/has_some_type.cpp
+    static_tests/is_primary_key_insertable.cpp
+    static_tests/is_column_with_insertable_primary_key.cpp
     sync_schema_tests.cpp
     tests.cpp
     tests2.cpp
@@ -45,29 +68,6 @@ add_executable(unit_tests
     prepared_statement_tests/insert_explicit.cpp
     pragma_tests.cpp
     simple_query.cpp
-    static_tests/is_printable.cpp
-    static_tests/is_bindable.cpp
-    static_tests/iterator_t.cpp
-    static_tests/arithmetic_operators_result_type.cpp
-    static_tests/tuple_conc.cpp
-    static_tests/node_tuple.cpp
-    static_tests/bindable_filter.cpp
-    static_tests/count_tuple.cpp
-    static_tests/member_traits_tests.cpp
-    static_tests/select_return_type.cpp
-    static_tests/column.cpp
-    static_tests/same_or_void.cpp
-    static_tests/foreign_key.cpp
-    static_tests/tuple_filter.cpp
-    static_tests/storage_traits.cpp
-    static_tests/tuple_helper.cpp
-    static_tests/function_static_tests.cpp
-    static_tests/aggregate_function_return_types.cpp
-    static_tests/core_function_return_types.cpp
-    static_tests/column_result_t.cpp
-    static_tests/has_some_type.cpp
-    static_tests/is_primary_key_insertable.cpp
-    static_tests/is_column_with_insertable_primary_key.cpp
     constraints/default.cpp
     constraints/unique.cpp
     constraints/foreign_key.cpp
@@ -114,6 +114,7 @@ add_executable(unit_tests
     unique_cases/join_iterator_ctor_compilation_error.cpp
     unique_cases/issue86.cpp
     unique_cases/issue937.cpp
+    unique_cases/issue663.cpp
     get_all_custom_containers.cpp
     select_constraints_tests.cpp
     backup_tests.cpp
@@ -123,7 +124,7 @@ add_executable(unit_tests
     row_id.cpp
     trigger_tests.cpp
     ast_iterator_tests.cpp
-    unique_cases/issue663.cpp
+    pointer_passing_interface.cpp
 )
 
 if(SQLITE_ORM_OMITS_CODECVT)
